+++ src/routes/users.route.js
@@ -1,5 +1,5 @@
 import { getAllDoctors } from "#controllers/users.controller";
-import { authenticationGuard } from "#middleware/authentication/authenticationGuard.middleware";
+import { authenticationGuard } from "src/middlewares/authentication/authenticationGuard.middleware";
 import { Router } from "express";
 
 const usersRouter = Router();