+++ src/routes/visits.routes.js
@@ -1,10 +1,11 @@
 import {
   acceptVisit,
   bookVisit,
+  cancelVisit,
   completeVisit,
   getAllVisits,
 } from "#controllers/visits.controller";
-import { authenticationGuard } from "#middleware/authentication/authenticationGuard.middleware";
+import { authenticationGuard } from "src/middlewares/authentication/authenticationGuard.middleware";
 import { Router } from "express";
 
 const visitsRouter = Router();
@@ -14,6 +15,7 @@ visitsRouter.use(authenticationGuard);
 visitsRouter.post("/book", bookVisit);
 visitsRouter.post("/complete", completeVisit);
 visitsRouter.post("/accept", acceptVisit);
+visitsRouter.post("/cancel", cancelVisit);
 visitsRouter.get("/", getAllVisits);
 
 export default visitsRouter;